Update perf tests and docs for batched GPU speedups
The main issue was the test pointing to an old legacy function which is no longer relevant. Once it points to the correct function when evaluating *single* sets of parameters and *single* bandpasses we are approx equal to sncosmo, depending on which gpu you use.

However, the above case is not really how jax bandflux shoudl be used. when integrated into gpu based samplers using jax, the entire process will be jitted and fused together allowing the sampler to evaluate many sets of parameters simultaneously and furthermore allow fast running of very large complex likelihoods, which is where the real speedup comes. We have updated the test to show this. When running on batched parameters we get an approx 100x speedup vs sncosmo. This will also enable even greater speedups when users are running complex likelihoods.

The test and documentation was immensely unclear, so I am very greatful that you picked this up so we can ammend for future users. I have also updated the paper, docs and tests to make this more clear. I also added some general optimisations as part of this PR which were not explicitly required to achieve the above, but useful nonetheless. We also show some more explicit tests in Leeney 2025 (now cited in paper)